Bug fixed. Particle stuck at two constraints.
The problem will end up in an infinite loop if a particle satisfies two or more
constraints at one place. That problem is due to the _getLinearHitTime method
which assumes that the reflective plane has an one to one relationship to the
constraint which yields K = 0 (consult the paper Eqn 2.26), but this is a poor
assumption because we can have two or more constraints that gives K = 0 at a
single place. Since the _getLinearHitTime may give incorrect hit times which
determine the next position of the particle, the particle may end up in a
location where the constraints are violated.

A simple fix will be to resample the velocity when stepping outside the feasible
space (i.e. new_b violates the constraints) because this scenario is uncommon,
but this strategy is not perfect as the velocity distribution should follow the
standard normal distribution. A better fix will be keeping track of the
reflective plane in a new variable, but this method might lead to the hit time
being 0 which implies that the particle is stuck on one place.